% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.oc.kb.R
\name{get.oc.kb}
\alias{get.oc.kb}
\title{Operating Characteristics for Single-agent Trials}
\usage{
get.oc.kb(
  target,
  p.true,
  ncohort,
  cohortsize,
  n.earlystop = 100,
  startdose = 1,
  marginL = 0.05,
  marginR = 0.05,
  cutoff.eli = 0.95,
  extrasafe = FALSE,
  offset = 0.05,
  ntrial = 1000
)
}
\arguments{
\item{target}{the target dose-limiting toxicity (DLT) rate.}

\item{p.true}{a vector containing the true toxicity probabilities of the
investigational dose levels.}

\item{ncohort}{a scalar specifying the total number of cohorts in the trial.}

\item{cohortsize}{the number of patients in the cohort.\cr
The default is 1.}

\item{n.earlystop}{the early stopping parameter. If the number of patients
treated at the current dose reaches \code{n.earlystop},
stop the trial and select the MTD based on the observed
data.\cr
The default value, 100, essentially turns off this type
of early stopping.}

\item{startdose}{the starting dose level for the trial.\cr
The default is 1.}

\item{marginL}{the difference between the target and the left bound of the
"target key" (proper dosing interval) to be defined.\cr
The default is 0.05.}

\item{marginR}{the difference between the target and the right bound of the
"target key" (proper dosing interval) to be defined.\cr
The default is 0.05.}

\item{cutoff.eli}{the cutoff to eliminate an overly toxic dose and all
higher doses for safety.\cr
The recommended value for general use and default is 0.95.}

\item{extrasafe}{set \code{extrasafe=TRUE} to impose a more stringent
stopping rule.\cr
The default is FALSE.}

\item{offset}{a small positive number (between 0 and 0.5) to control how
strict the stopping rule is when \code{extrasafe=TRUE}. A
larger value leads to a stricter stopping rule.\cr
The default value of 0.05 generally works well.}

\item{ntrial}{the total number of trials to be simulated. \cr
The default value is 1000.}
}
\value{
The function returns the operating characteristics of the KEYBOARD
  design (single-agent) as a list, which includes: \cr
\enumerate{
  \item selection percentage at each dose level (\code{$selpercent}),\cr
  \item the average number of patients treated at each dose level (\code{$npatients}),\cr
  \item the average number of toxicities observed at each dose level (\code{$ntox}),\cr
  \item the average number of toxicities (\code{$totaltox}),\cr
  \item the average number of patients (\code{$totaln}),\cr
  \item the percentage of early stopping due to toxicity without selecting the MTD (\code{$percentstop}),\cr
  \item risk of overdosing 60\% or more of patients (\code{$overdose60}),\cr
  \item risk of overdosing 80\% or more of patients (\code{$overdose80}),\cr
  \item a data.frame object containing simulation parameters, such as \code{target, p.true,} etc. (\code{$simu.setup}).
}
}
\description{
Generates the operating characteristics of the KEYBOARD design for
single-agent trials.
}
\details{
The operating characteristics of the KEYBOARD design are generated by
simulating trials under the pre-specified true toxicity probabilities of the
investigational doses.

The KEYBOARD design defines a series of equal-width dosing intervals (or
"keys") to present the potential locations of the true toxicity of a dose
and guide the dose escalation and de-escalation between cohorts of patients.
The KEYBOARD design starts by specifying a proper dosing interval, referred
to as the "target key", based on which it forms a series of equally wide
keys that span the remaining of the range from 0 to 1.

The KEYBOARD design relies on the posterior distribution of the toxicity
probability to guide dosage. To make the decision of dose escalation and
de-escalation, given the observed data at the current dose, we identify the
key that has the highest posterior probability, which we refer to as the
"strongest key". This key represents where the true dose-limiting toxicity
(DLT) rate of the current dose is most likely located. If the strongest key
is located on the left side of the target key, we escalate the dose (because
it means that the observed data suggests that the current dose is most
likely to represent under-dosing); if the strongest key is located on the
right side of the target key, we de-escalate the dose (because the data
suggests that the current dose represents overdosing); and if the strongest
key is the target key, we retain the current dose (because the observed data
supports that the current dose is most likely to be in the proper dosing
interval).
Graphically, the strongest key is the one with the largest area under the
posterior distribution curve of the DLT rate of the current dose.

An attractive feature of the KEYBOARD design is that its dose escalation and
de-escalation rule can be tabulated before the onset of the trial. Thus,
when conducting the trial, no calculation or model fitting is needed, and we
only need to count the number of DLTs observed at the current dose and make
the decision of dose escalation and de-escalation based on the pre-tabulated
decision rules.

Given all observed data, the KEYBOARD design uses a statistical technique
called isotonic regression to obtain an efficient statistical estimate of
the maximum tolerated dose (MTD) by utilizing the fact that toxicity
presumably increases with the dose.

For patient safety, the following dose elimination rule is evaluated after
each cohort:
if at least 3 patients have been treated at the given dose and
the observed data indicate that there is more than 95% chance that the
current dose is above the maximum tolerated dose (MTD), we eliminate that
and higher doses from the trial to prevent exposing future patients to these
overly toxic doses. The probability threshold for elimination can be
specified with \code{cutoff.eli}. When a dose is eliminated, the design
recommends the next lower dose for treating the next patient.

The KEYBOARD design has two built-in stopping rules:\cr
(1) stop the trial if the lowest dose is eliminated due to toxicity, and no
    dose should be selected as the MTD; and \cr
(2) stop the trial and select the MTD if the number of patients treated at
    the current dose reaches \code{n.earlystop}.

The first stopping rule is a safety rule to protect patients from the case
in which all doses are overly toxic. The rationale for the second stopping
rule is that when there is a large number (i.e., \code{n.earlystop}) of
patients assigned to a dose, it means that the dose-finding algorithm has
approximately converged. Thus, we can stop the trial early and select the
MTD to save sample size and reduce the trial duration.

For some applications, investigators may prefer a stricter stopping rule
than rule (1) for extra safety when the lowest dose is overly toxic.
This can be achieved by setting \code{extrasafe=TRUE}, which imposes the
following, stricter, safety stopping rule:\cr
Stop the trial if \cr
(i) the number of patients treated at the lowest dose \eqn{\ge 3}, and \cr
(ii) \deqn{Pr((toxicity rate of the lowest dose > target) | data)
           > cutoff.eli - offset}
As a tradeoff, the strong stopping rule will decrease the MTD selection
percentage when the lowest dose actually is the MTD.
}
\section{Uses}{

This function uses \code{\link{get.boundary.kb}} and
\code{\link{select.mtd.kb}}.
}

\examples{
### Single-agent trial ###

oc <- get.oc.kb(target=0.3, p.true=c(0.05, 0.15, 0.3, 0.45, 0.6),
                ncohort=20, cohortsize=3, ntrial=1000)

oc

summary.kb(oc)

plot.kb(oc$ntox)  # can plot either $selpercent, $npatients, or $ntox

}
\references{
Yan F, Mandrekar SJ, Yuan Y. KEYBOARD: A Novel Bayesian Toxicity Probability
Interval Design for Phase I Clinical Trials.
\emph{Clinical Cancer Research}. 2017; 23:3994-4003.
http://clincancerres.aacrjournals.org/content/23/15/3994.full-text.pdf
}
\seealso{
Other single-agent functions: 
\code{\link{get.boundary.kb}()},
\code{\link{select.mtd.kb}()}
}
\concept{single-agent functions}
