% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.oc.obd2.kb.R
\name{get.oc.obd2.kb}
\alias{get.oc.obd2.kb}
\title{Generate Operating Characteristics for OBD Finding}
\usage{
get.oc.obd2.kb(
  target.toxicity,
  target.efficacy,
  ncohort = 10,
  cohortsize = 3,
  n.early = 100,
  startdose = 1,
  p.true,
  q.true,
  ntrial = 1000,
  seed = 6,
  p1 = 0.15,
  p2 = 0.4,
  q1 = 0.3,
  q2 = 0.6,
  cutoff.eli.toxicity = 0.95,
  cutoff.eli.efficacy = 0.3,
  w1.toxicity = 0.33,
  w2.toxicity = 1.09,
  indicator = target.toxicity
)
}
\arguments{
\item{target.toxicity}{the target toxicity rate}

\item{target.efficacy}{the target efficacy rate}

\item{ncohort}{the total number of cohorts}

\item{cohortsize}{the cohortsize}

\item{n.early}{the early stopping parameter. If the number of patients treated at
the current dose reaches \code{n.early}, stop the trial
and select the optimal biological dose (OBD) based on the observed data. The default
value \code{n.early=100} essentially turns off the type
of early stopping.}

\item{startdose}{the starting dose level}

\item{p.true}{a vector containing the true toxicity probabilities of the
investigational dose levels.}

\item{q.true}{a vector containing the true efficacy  probabilities of the
investigational dose levels.}

\item{ntrial}{the total number of trials to be simulated}

\item{seed}{the random seed for simulation}

\item{p1}{the cutoff lower limit for safety utility function 1, described in the Details section.}

\item{p2}{the cutoff upper limit for safety utility function 1, described in the Details section.}

\item{q1}{the cutoff lower limit for efficacy utility function 1, described in the Details section.}

\item{q2}{the cutoff upper limit for efficacy utility function 1, described in the Details section.}

\item{cutoff.eli.toxicity}{the cutoff to eliminate a dose with unacceptable high toxicity for safety.
The recommended value is (\code{cutoff.eli.toxicity=0.95}).}

\item{cutoff.eli.efficacy}{the cutoff for the futility rule, the acceptable lowest efficacy.
The recommended value is  (\code{cutoff.eli.toxicity=0.3})}

\item{w1.toxicity}{the weight for toxicity utility function 2 and 3, described in the Details section.}

\item{w2.toxicity}{the weight for toxicity utility function 3, described in the Details section.}

\item{indicator}{the indicator cutoff for utility function 3, described in the Details section.}
}
\value{
\code{get.oc.obd2.kb()} returns the operating characteristics of the KEYBOARD design as a list, including:\cr
\enumerate{
   \item selection percentage at each dose level using utility function 1 (\code{$selpercent1}), \cr
   \item selection percentage at each dose level using utility function 2 (\code{$selpercent2}), \cr
   \item selection percentage at each dose level using utility function 3 (\code{$selpercent3}), \cr
   \item the number of patients treated at each dose level (\code{$npatients}), \cr
   \item the number of dose limiting toxicities (DLTs) observed at each dose level (\code{$ntox}), \cr
   \item the number of responders observed at each dose level (\code{$neff}), \cr
   \item the average number of DLTs (\code{$totaltox}), \cr
   \item the average number of responders (\code{$totaleff}), \cr
   \item the average number of patients (\code{$totaln}), \cr
   \item the percentage of early stopping without selecting the OBD using utility function 1 (\code{$percentstop1}), \cr
   \item the percentage of early stopping without selecting the OBD using utility function 2 (\code{$percentstop2}), \cr
   \item the percentage of early stopping without selecting the OBD using utility function 3 (\code{$percentstop3}), \cr
   \item data.frame (\code{$simu.setup}) containing simulation parameters, such as target, p.true, etc.


}
}
\description{
Report operating characteristics for optimal biological dose (OBD) finding
}
\details{
A large number of  trials are simulated to characterize the operating characteristics of the KEYBOARD design under the prespecified true toxicity probabilities and true efficacy probabilities of the investigational doses. Dose assignment rule follows the rule described in the function \code{get.decision.obd.kb()}.

There are stopping rules the KEYBOARD design:
\enumerate{
\item stop the trial if the lowest dose is eliminated from the trial due to high unacceptable toxicity.
\item stop the trial if the number of patients treated at current dose is larger than or equal to \code{n.earlystop}.
}
}
\examples{
target.toxicity <- 0.30
target.efficacy <- 0.40
p.true <-c(0.08,0.30,0.60,0.80)
q.true <- c(0.25,0.40,0.25,0.50)
oc.obd2.kb <- get.oc.obd2.kb(target.toxicity=target.toxicity,
              target.efficacy= target.efficacy, ncohort=20,
              cohortsize= 3,  p.true= p.true, q.true= q.true)
oc.obd2.kb
summary.kb(oc.obd2.kb)
plot.kb(oc.obd2.kb)
plot.kb(oc.obd2.kb$selpercent1)
plot.kb(oc.obd2.kb$selpercent2)
plot.kb(oc.obd2.kb$selpercent3)
plot.kb(oc.obd2.kb$npatients)
plot.kb(oc.obd2.kb$ntox)
plot.kb(oc.obd2.kb$neff)

}
\references{
Daniel H. Li, James B. Whitmore, Wentian Guo and Yuan Ji  Toxicity and Efficacy Probability Interval Design for Phase I Adoptive Cell Therapy Dose-Finding Clinical Trials
\emph{Clinical Cancer Research}. 2017; 23:13-20.
https://clincancerres.aacrjournals.org/content/23/1/13.long

Suyu Liu, Valen E. Johnson. A robust Bayesian dose-finding design for phase I/II clinical trials
\emph{Biostatistics}. 2016; 17:249-263.
https://academic.oup.com/biostatistics/article/17/2/249/1744018

Yanhong Zhou, J. Jack Lee, Ying Yuan. A utility-based Bayesian optimal interval (U-BOIN) phase I/II design to identify the optimal biological dose for targeted and immune therapies
\emph{Statistics in Medicine}. 2019; 38:S5299-5316.
https://onlinelibrary.wiley.com/doi/epdf/10.1002/sim.8361
}
\seealso{
Other single-agent phase I/II functions: 
\code{\link{get.decision.obd.kb}()},
\code{\link{get.decision.obd2.kb}()},
\code{\link{get.oc.obd.kb}()},
\code{\link{select.obd.kb}()}
}
\concept{single-agent phase I/II functions}
