% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.kb.R
\name{summary.kb}
\alias{summary.kb}
\title{Summarize Objects Generated by the KEYBOARD Design Package}
\usage{
\method{summary}{kb}(object, ...)
}
\arguments{
\item{object}{the object (returned by another function) to be described.}

\item{...}{ignored.}
}
\value{
This function is helpful to interpret the objects returned by other
  KEYBOARD package functions.
}
\description{
Unpacks objects returned by other functions with descriptive descriptions of
their results. The following functions are supported:\cr
(1) get.oc.kb(), get.oc.comb.kb(), get.oc.obd.kb(), and get.oc.obd2.kb(), which yield the operating
    characteristics of trials simulated using the KEYBOARD design.\cr
(2) select.mtd.kb() and select.mtd.comb.kb(), which yield the MTD and other
    stats.\cr
(3) next.comb.kb(), which indicates the dose combination to administer to
    the next cohort. \cr
(4) select.obd.kb(), which yields the OBD.
}
\examples{
### Single-agent trial ###

## Summarize the object returned by get.oc.kb()
oc.single <- get.oc.kb(target=0.3, p.true=c(0.05, 0.15, 0.3, 0.45, 0.6),
                       ncohort=10, cohortsize=3, ntrial=1000)
summary.kb(oc.single)


## Summarize the object returned by select.mtd.kb()
n <- c(3, 3, 15, 9, 0)
y <- c(0, 0, 4, 4, 0)
sel.single <- select.mtd.kb(target=0.3, npts=n, ntox=y)
summary.kb(sel.single)


### Drug-combination trial ###

## Summarize the object returned by get.oc.comb.kb()
p.true <- matrix(c(0.02, 0.04, 0.08, 0.14,
                   0.08, 0.25, 0.42, 0.48,
                   0.25, 0.45, 0.50, 0.60), byrow=TRUE, ncol=4)
oc.comb <- get.oc.comb.kb(target=0.25, p.true, ncohort=16, cohortsize=3,
                          ntrial=100)
summary.kb(oc.comb)


## Summarize the object returned by select.mtd.comb.kb()
n <- matrix(c(6, 3, 0, 0,
              6, 24, 9, 0,
              0, 0, 0, 0), ncol=4, byrow=TRUE)
y <- matrix(c(0, 0, 0, 0,
              1, 5, 4, 0,
              0, 0, 0, 0), ncol=4, byrow=TRUE)
sel.comb <- select.mtd.comb.kb(target=0.25, npts=n, ntox=y)
summary.kb(sel.comb)


## Summarize the object returned by next.comb.kb()
n <- matrix(c(3, 0, 0, 0,
              0, 0, 0, 0,
              0, 0, 0, 0), ncol=4, byrow=TRUE)
y <- matrix(c(0, 0, 0, 0,
              0, 0, 0, 0,
              0, 0, 0, 0), ncol=4, byrow=TRUE)
nxt.comb <- next.comb.kb(target=0.25, npts=n, ntox=y, dose.curr=c(1, 1))
summary.kb(nxt.comb)

## get.oc.obd
toxicity.low <- 0.15
toxicity.moderate <- 0.33
toxicity.high <- 0.40
efficacy.low <- 0.20
efficacy.moderate <- 0.40
efficacy.high <- 0.60
target.toxicity <- 0.30
target.efficacy <- 0.40
p.true <-c(0.08,0.30,0.60,0.80)
q.true <- c(0.25,0.40,0.25,0.50)
oc.obd.kb <- get.oc.obd.kb(toxicity.low = toxicity.low,
             toxicity.moderate= toxicity.moderate,
             toxicity.high = toxicity.high,
             efficacy.low = efficacy.low,
             efficacy.moderate = efficacy.moderate,
             efficacy.high = efficacy.high,
             target.toxicity=target.toxicity,
             target.efficacy= target.efficacy,
             p.true= p.true, q.true= q.true)
summary.kb(oc.obd.kb)

## OBD selection
target.toxicity=0.3
target.efficacy=0.4
npts <- c(3,6,12,3,3)
ntox <-  c(1,2,4,2,3)
neff <-  c(0,0,5,1,1)
obd <- select.obd.kb (target.toxicity=target.toxicity,
       target.efficacy= target.efficacy, npts = npts,
       ntox = ntox, neff =  neff)
summary.kb(obd)

}
\references{

}
\seealso{
\code{\link{plot.kb}}
}
